pyproject.toml
[project]
name = "continue-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "lancedb>=0.16.1",
    "numpy>=2.2.5",
    "openai>=1.76.0",
    "pandas>=2.2.3",
    "pylance>=0.26.0",
    "pysocks>=1.7.1",
    "python-dotenv>=1.1.0",
    "rake-nltk>=1.0.6",
    "requests>=2.32.3",
    "socksio>=1.0.0",
    "tantivy",
    "tdqm>=0.0.1",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.2",
]
